+++ jgenesis-common/README.md
@@ -2,6 +2,8 @@
 
 Common code used across most of the other crates.
 
-`jgenesis_traits::num` contains traits useful for accessing specific bits in primitive integer values.
+`jgenesis_common::audio` contains core audio resampling code.
 
-`jgenesis_traits::frontend` contains traits for interaction between the emulation cores and emulation frontends, such as an emulator trait, a renderer trait, and an audio output trait.
+`jgenesis_common::num` contains traits useful for accessing specific bits in primitive integer values.
+
+`jgenesis_common::frontend` contains traits for interaction between the emulation cores and emulation frontends, such as an emulator trait, a renderer trait, and an audio output trait.
