+++ jgenesis-common/src/lib.rs
@@ -1,2 +1,3 @@
+pub mod audio;
 pub mod frontend;
 pub mod num;
